Add JsonService and replace Newtonsoft.Json
Created and implemented a new JsonService to handle all JSON serialization and deserialization throughout the application. This included the removal of all usage of Newtonsoft.Json and replacing all instances with System.Text.Json. The application's performance was improved by this change. The documentation was also updated to reflect the changes.